+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Method to enable converting the stagedElements to drafts
 - Method to delete an import by id
 - Method to update the members of dataElementGroup/record
+- Method to return all available paths for a given element
 ### Changed
 - Separate namespaceService from elementService
 ### Deprecated